requirements.txt
azure-cosmosdb-table
azure-functions
matplotlib
pandas
pandas_profiling
requests